+++ packages/types/lib/change-case/camel-to-kebab-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import CamelToKebabCase from "./camel-to-kebab-case";
 
 type CamelToKebabCaseKeys<T extends Json, D extends number = -1> = [D] extends [
