+++ packages/types/lib/change-case/camel-to-snake-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import CamelToSnakeCase from "./camel-to-snake-case";
 
 type CamelToSnakeCaseKeys<T extends Json, D extends number = -1> = [D] extends [
