+++ packages/types/lib/change-case/kebab-to-pascal-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import KebabToPascalCase from "./kebab-to-pascal-case";
 
 type KebabToPascalCaseKeys<T extends Json, D extends number = -1> = [
