+++ packages/types/lib/change-case/kebab-to-snake-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import KebabToSnakeCase from "./kebab-to-snake-case";
 
 type KebabToSnakeCaseKeys<T extends Json, D extends number = -1> = [D] extends [
