+++ packages/types/lib/change-case/pascal-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import PascalCase from "./pascal-case";
 
 type PascalCaseKeys<T extends Json, D extends number = -1> = [D] extends [0]
