+++ packages/types/lib/change-case/pascal-to-kebab-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import PascalToKebabCase from "./pascal-to-kebab-case";
 
 type PascalToKebabCaseKeys<T extends Json, D extends number = -1> = [
