+++ packages/types/lib/change-case/pascal-to-snake-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import PascalToSnakeCase from "./pascal-to-snake-case";
 
 type PascalToSnakeCaseKeys<T extends Json, D extends number = -1> = [
