+++ packages/types/lib/change-case/snake-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import SnakeCase from "./snake-case";
 
 type SnakeCaseKeys<T extends Json, D extends number = -1> = [D] extends [0]
