+++ packages/types/lib/change-case/snake-to-camel-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import SnakeToCamelCase from "./snake-to-camel-case";
 
 type SnakeToCamelCaseKeys<T extends Json, D extends number = -1> = [D] extends [
