+++ packages/types/lib/change-case/snake-to-kebab-case-keys.ts
@@ -1,5 +1,5 @@
-import Json, { JsonObject } from "../json";
-import Prev from "../prev";
+import Json, { JsonObject } from "../core/json";
+import Prev from "../core/prev";
 import SnakeToKebabCase from "./snake-to-kebab-case";
 
 type SnakeToKebabCaseKeys<T extends Json, D extends number = -1> = [D] extends [
