+++ packages/types/lib/core/nil.ts
