+++ packages/types/lib/core/paths.ts
