+++ packages/types/lib/core/type.ts
