+++ packages/types/lib/index.ts
@@ -1,10 +1,2 @@
-export { default as Type } from "./type";
-export { default as DeepPartial } from "./deep-partial";
-export { default as Nullish } from "./nullish";
-export { default as Json, JsonObject, JsonArray } from "./json";
-export { default as Join } from "./join";
-export { default as Prev } from "./prev";
-export { default as Paths } from "./paths";
-export { default as Leaves } from "./leaves";
-
 export * from "./change-case";
+export * from "./core";
